requirements.txt
django-pytest==0.2.0
